Move `TxOut` type into separate module.
This allows the `TxOut` type to be imported without having to depend on
everything in the `Tx.Tx` module.

Note that this commit creates an orphan `Buildable` instance: this will
be fixed in a later commit.
